Update service raspberrypi - auto-start, Update Documentation
