+++ pi_server/documentation/doc.md
@@ -2,15 +2,15 @@
 #Configuration Raspberrypi:
 
 ##Creation service in /lib/systemd/system
-+ pi@raspberrypi:/lib/systemd/system $ sudo nano kibblepi_start_server.service
++ pi@raspberrypi:/lib/systemd/system $ sudo nano kibblepi_server.service
 
 ##Start the service manually
-+ sudo systemctl start kibblepi_start_server.service
++ sudo systemctl start kibblepi_server.service
 ##Stop the service manually
-+ sudo systemctl stop kibblepi_start_server.service
++ sudo systemctl stop kibblepi_server.service
 
 ##Enable the service automatically
-+ sudo systemctl enable kibblepi_start_server.service
++ sudo systemctl enable kibblepi_server.service
 
-##Enable the service automatically
-+ sudo systemctl enable kibblepi_start_server.service
+##Disable the service automatically
++ sudo systemctl disable kibblepi_server.service
