+++ pi_server/documentation/kibblepi_server.service
@@ -3,7 +3,7 @@ Description= Kibblepi - Web server - automatic start
 After=multi-use.target
 
 [Service]
-ExecStart=/usr/bin/python3 /home/pi/Documents/python_raspberry/web_server_flask_1.py
+ExecStart=/usr/bin/python3 /home/pi/kibblepi/pi_server/pi_server/__main__.py
 User=pi
 
 [Install]
