+++ pi_server/pi_server/views/__init__.py
@@ -1,6 +1,2 @@
-from pi_server import app
 
 
-@app.route("/")
-def index():
-    return "Hello from Raspberry PI"
